Fix stats reporting when using >3 byte UID's
The '1' defaults to an int, left-shifting it for >3 byte UID's ends up with 0.

+++ src/uid/UniqueId.java
@@ -161,7 +161,7 @@ public final class UniqueId implements UniqueIdInterface {
   
   /** The largest possible ID given the number of bytes the IDs are represented on. */
   public long maxPossibleId() {
-    return (1 << id_width * Byte.SIZE) - 1;
+    return ((long) 1 << id_width * Byte.SIZE) - 1;
   }
   
   /**
